+++ smcsmc/__init__.py
@@ -1,5 +1,5 @@
 from .model import Smcsmc
-from .square_model import *
+from .simulations import *
 from .populationmodels import Population
 from .plot import *
 from .utils import ts_to_seg, run_smcsmc, vcf_to_seg
